Support multiple vcpus on a cpu.
You can use command "vcpu_run_ext" to create more than one vcpus.
Its usage is the same as vcpu_run, and every time "vcpu_run_ext"
is called, a vcpu will be created and run on this cpu.

The virt subsystem of gos will allocate vmid itself, and the log of
vm has a prefix of "[VM%d log]"(%d is vmid of this vm).

Such as:
[VM1 log] XXXXX —— the log of vm1
[VM2 log] XXXXX —— the log of vm2

Ps. The usage of vcpu_run remains unchanged. It only creates a vcpu
when there is no existing vcpu in the cpu, and it will return the
first vcpu when there is a vcpu in the cpu.

Signed-off-by: YaXing Guo <guoyaxing@bosc.ac.cn>