Add boot_option in ctx in mysbi to configure auto run options.
Remove the original auto_run.S which is used to include auto_run.bin.

Instead, pass the boot_option where its address is fixed(0x801f0000)
in mysbi.lds to gos and use the content in it as auto run command.

In pack.sh, pack the auto_run.bin to 0x801f0000. So,the content in
auto_run.bin is still as the auto run command for gos.

In addition, some methods can be used to modify the content in 0x801f0000
as the auto run command for gos. Such as, through fpga jtag during programing.

Signed-off-by: YaXing Guo <guoyaxing@bosc.ac.cn>